+++ subsurface/__init__.py
@@ -1,2 +1,1 @@
 from .curve import *
-from .seismic import *
