--- subsurface/io/__init__.py
@@ -1,1 +1,0 @@
-from .seismic import *
